requirements.txt
# Paquetes esenciales para Flask con versiones compatibles
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
blinker==1.6.2

# Autenticación y base de datos
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
WTForms==3.0.1

# Procesamiento de API y OpenAI
openai==1.72.0
python-dotenv==1.0.0
requests==2.28.2

# Utilidades
tqdm==4.65.0